Extract remote-build listener lifecycle from device_builder.py
device_builder.py was the largest module in the package (1379 lines).
Move the remote-build peer-link listener lifecycle — bind / teardown /
rebuild of the Noise-XX receiver listener and its mDNS advertise — into
a _RemoteBuildLifecycleMixin in a new esphome_device_builder/
_remote_build_lifecycle.py, mixed into DeviceBuilder. Behaviour-free
refactor: the runner reference and lifecycle lock stay owned by
DeviceBuilder.__init__; every method keeps its signature, docstring, and
public name, so the API surface and instance-patch test seams are
unchanged. device_builder.py drops to 976 lines.

A mixin (not a free-function delegate module) keeps every access as
self._x, so the cross-class SLF001 protection the project deliberately
preserves on top-level files is untouched — no lint-config change.

The _strip_server_header_middleware moved with it; the one test that
imported it and the moved resolve_bind_host monkeypatch / caplog logger
seams are repointed to the new module.
